[refactor]: designsystem modulation
