+++ app/src/main/java/org/sopt/dosoptjaewon/domain/model/User.kt
