+++ core/designsystem/src/main/res/values/colors.xml
@@ -1,9 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <color name="purple_200">#FFBB86FC</color>
+    <color name="purple_500">#FF6200EE</color>
+    <color name="purple_700">#FF3700B3</color>
+    <color name="teal_200">#FF03DAC5</color>
+    <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
 
-    <!-- Pophory color 가져옴 ㅎㅎ-->
     <color name="pophory_purple">#725BFF</color>
     <color name="pophory_red">#FC4646</color>
     <color name="pophory_black">#FF212121</color>
