+++ core/designsystem/src/main/res/values/themes.xml
@@ -14,7 +14,8 @@
         <item name="textAppearanceBodyLarge">@style/TextAppearance.DoSoptJaewon.HeadLineBold</item>
         <item name="textAppearanceHeadlineLarge">@style/TextAppearance.DoSoptJaewon.HeadLineLarge</item>
         <item name="textAppearanceHeadlineMedium">@style/TextAppearance.DoSoptJaewon.HeadLineMedium</item>
-        <item name="textAppearanceHeadlineSmall">@style/TextAppearance.DoSoptJaewon.HeadLineSmall</item>
+        <item name="textAppearanceHeadlineSmall">@style/TextAppearance.DoSoptJaewon.HeadLineSmall
+        </item>
     </style>
 
     <style name="Theme.DoSoptJaewon" parent="Base.Theme.DoSoptJaewon" />
