+++ settings.gradle
@@ -15,3 +15,4 @@ dependencyResolutionManagement {
 rootProject.name = "DoSoptJaewon"
 include ':app'
 include ':core:common'
+include ':core:designsystem'
